ARMCMx: rules_code.ld: align .vectors to 256
See the comment in the commit for more details.

Alignment generally doesn’t matter when the .vectors section is the first one
placed on the flash, ending up at address 0.

The Teensy 4.1, however, requires flashconfig to come first (more on that in
another commit to ChibiOS-Contrib), and isn’t happy with un-aligned vector
tables. In my case, it rebooted after enabling the SysTick_Handler interrupt.